Move TextClickable composable into "designsystem" package.
