+++ app/src/main/java/nerd/tuxmobil/fahrplan/congress/about/AboutComposables.kt
@@ -22,7 +22,6 @@ import androidx.compose.ui.unit.dp
 import nerd.tuxmobil.fahrplan.congress.BuildConfig
 import nerd.tuxmobil.fahrplan.congress.R
 import nerd.tuxmobil.fahrplan.congress.about.AboutViewEvent.OnPostalAddressClick
-import nerd.tuxmobil.fahrplan.congress.commons.ClickableText
 import nerd.tuxmobil.fahrplan.congress.commons.MultiDevicePreview
 import nerd.tuxmobil.fahrplan.congress.commons.TextResource
 import nerd.tuxmobil.fahrplan.congress.commons.TextResource.Empty
@@ -31,6 +30,7 @@ import nerd.tuxmobil.fahrplan.congress.commons.TextResource.PostalAddress
 import nerd.tuxmobil.fahrplan.congress.designsystem.dividers.DividerHorizontal
 import nerd.tuxmobil.fahrplan.congress.designsystem.templates.Scaffold
 import nerd.tuxmobil.fahrplan.congress.designsystem.texts.Text
+import nerd.tuxmobil.fahrplan.congress.designsystem.texts.TextClickable
 import nerd.tuxmobil.fahrplan.congress.designsystem.themes.EventFahrplanTheme
 import nerd.tuxmobil.fahrplan.congress.extensions.toTextUnit
 
@@ -217,7 +217,7 @@ private fun AboutClickableText(
     textAlign: TextAlign = TextAlign.Start,
     onClick: (String) -> Unit = {},
 ) {
-    ClickableText(
+    TextClickable(
         textResource = textResource,
         fontSize = dimensionResource(R.dimen.about_text).toTextUnit(), // To match font size of AboutText
         textAlign = textAlign,
